pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "metaquery-actuary"
version = "0.1.0"
description = "Field-level SQL governance for model validation & auditability"
requires-python = ">=3.11"
dependencies = [
  "pyyaml>=6.0",
  "typer>=0.9.0",
  "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4.0",
  "ruff>=0.1.0",
]

[project.scripts]
metaquery = "metaquery.cli:app"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
